+++ docs/changelog.md
@@ -2,9 +2,14 @@
 
 All notable changes to this project will be documented below.
 
+
 #### geospatial.analyze.coverage
 
 * v0.1dev: **geospatial.analyze.coverage**(*img, bin_mask, geojson*)
+
+#### geospatial.points2roi_circle
+
+* v0.1dev: rois = **geospatial.points2roi_circle**(*img, geojson, radius*)
 
 #### geospatial.points_to_geojson